+++ test_desktop/main.cpp
@@ -52,7 +52,7 @@ public:
         return 0;
     }
 
-    size_t getFrameSize() const override {
+    int getFrameSize() const override {
         return size_;
     }
